resolved Created dataAdd.js and clean.sh fix for Linux/MacOS differences

+++ bin/clean.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
 # find and delete backup files
-#find-bak.sh -exec rm {} \;
-find-bak.sh -delete
+if which sw_vers > /dev/null 2>&1 ; then
+	# MACOS!
+	find-bak.sh -delete
+else
+	find-bak.sh -exec rm {} \;
+fi
